// test/integration/test/sql-orm-client/fixtures/polymorphism/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Account: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Admin: { readonly role: CodecTypes['pg/text@1']['output'] };
    readonly Bug: {
      readonly assigneeId: CodecTypes['pg/int4@1']['output'] | null;
      readonly severity: CodecTypes['pg/text@1']['output'];
    };
    readonly Epic: { readonly scope: CodecTypes['pg/text@1']['output'] };
    readonly Feature: {
      readonly assigneeId: CodecTypes['pg/int4@1']['output'] | null;
      readonly priority: CodecTypes['pg/int4@1']['output'];
    };
    readonly Person: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Project: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Regular: { readonly plan: CodecTypes['pg/text@1']['output'] | null };
    readonly Task: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly projectId: CodecTypes['pg/int4@1']['output'] | null;
      readonly reporterId: CodecTypes['pg/int4@1']['output'] | null;
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly type: CodecTypes['pg/text@1']['output'];
    };
    readonly TaskComment: {
      readonly body: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly taskId: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly Ticket: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly ownerId: CodecTypes['pg/int4@1']['output'] | null;
      readonly subject: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly accountId: CodecTypes['pg/int4@1']['output'] | null;
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly kind: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Account: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Admin: { readonly role: CodecTypes['pg/text@1']['input'] };
    readonly Bug: {
      readonly assigneeId: CodecTypes['pg/int4@1']['input'] | null;
      readonly severity: CodecTypes['pg/text@1']['input'];
    };
    readonly Epic: { readonly scope: CodecTypes['pg/text@1']['input'] };
    readonly Feature: {
      readonly assigneeId: CodecTypes['pg/int4@1']['input'] | null;
      readonly priority: CodecTypes['pg/int4@1']['input'];
    };
    readonly Person: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Project: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Regular: { readonly plan: CodecTypes['pg/text@1']['input'] | null };
    readonly Task: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly projectId: CodecTypes['pg/int4@1']['input'] | null;
      readonly reporterId: CodecTypes['pg/int4@1']['input'] | null;
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly type: CodecTypes['pg/text@1']['input'];
    };
    readonly TaskComment: {
      readonly body: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly taskId: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly Ticket: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly ownerId: CodecTypes['pg/int4@1']['input'] | null;
      readonly subject: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly accountId: CodecTypes['pg/int4@1']['input'] | null;
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly kind: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly accounts: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly epics: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly scope: CodecTypes['pg/text@1']['output'];
    };
    readonly features: {
      readonly feature_assignee_person_id: CodecTypes['pg/int4@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly priority: CodecTypes['pg/int4@1']['output'];
    };
    readonly people: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly projects_tbl: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly task_comments: {
      readonly body: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly task_id: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly tasks: {
      readonly bug_assignee_person_id: CodecTypes['pg/int4@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly project_id: CodecTypes['pg/int4@1']['output'] | null;
      readonly reporter_id: CodecTypes['pg/int4@1']['output'] | null;
      readonly severity: CodecTypes['pg/text@1']['output'] | null;
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly type: CodecTypes['pg/text@1']['output'];
    };
    readonly tickets: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly owner_id: CodecTypes['pg/int4@1']['output'] | null;
      readonly subject: CodecTypes['pg/text@1']['output'];
    };
    readonly users: {
      readonly account_id: CodecTypes['pg/int4@1']['output'] | null;
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly kind: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly plan: CodecTypes['pg/text@1']['output'] | null;
      readonly role: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
// ...
export namespace Models {
  export type public_Account = {
    id: CodecTypes['pg/int4@1']['output'];
    name: CodecTypes['pg/text@1']['output'];
    members: public_AnyUser[];
    readonly [RelationKeys]?: 'members';
  };
  export type public_Admin = {
    accountId: CodecTypes['pg/int4@1']['output'] | null;
    email: CodecTypes['pg/text@1']['output'];
    id: CodecTypes['pg/int4@1']['output'];
    kind: 'admin';
    name: CodecTypes['pg/text@1']['output'];
    role: CodecTypes['pg/text@1']['output'];
    account: public_Account | null;
    readonly [RelationKeys]?: 'account';
  };
  export type public_Bug = {
    id: CodecTypes['pg/int4@1']['output'];
    projectId: CodecTypes['pg/int4@1']['output'] | null;
    reporterId: CodecTypes['pg/int4@1']['output'] | null;
    title: CodecTypes['pg/text@1']['output'];
    type: 'bug';
    assigneeId: CodecTypes['pg/int4@1']['output'] | null;
    severity: CodecTypes['pg/text@1']['output'];
    comments: public_TaskComment[];
    project: public_Project | null;
    reporter: public_Person | null;
    assignee: public_Person | null;
    readonly [RelationKeys]?: 'comments' | 'project' | 'reporter' | 'assignee';
  };
  export type public_Epic = {
    id: CodecTypes['pg/int4@1']['output'];
    projectId: CodecTypes['pg/int4@1']['output'] | null;
    reporterId: CodecTypes['pg/int4@1']['output'] | null;
    title: CodecTypes['pg/text@1']['output'];
    type: 'epic';
    scope: CodecTypes['pg/text@1']['output'];
    comments: public_TaskComment[];
    project: public_Project | null;
    reporter: public_Person | null;
    readonly [RelationKeys]?: 'comments' | 'project' | 'reporter';
  };
  export type public_Feature = {
    id: CodecTypes['pg/int4@1']['output'];
    projectId: CodecTypes['pg/int4@1']['output'] | null;
    reporterId: CodecTypes['pg/int4@1']['output'] | null;
    title: CodecTypes['pg/text@1']['output'];
    type: 'feature';
    assigneeId: CodecTypes['pg/int4@1']['output'] | null;
    priority: CodecTypes['pg/int4@1']['output'];
    comments: public_TaskComment[];
    project: public_Project | null;
    reporter: public_Person | null;
    assignee: public_Person | null;
    readonly [RelationKeys]?: 'comments' | 'project' | 'reporter' | 'assignee';
  };
  export type public_Person = {
    id: CodecTypes['pg/int4@1']['output'];
    name: CodecTypes['pg/text@1']['output'];
    tasks: public_AnyTask[];
    readonly [RelationKeys]?: 'tasks';
  };
  export type public_Project = {
    id: CodecTypes['pg/int4@1']['output'];
    name: CodecTypes['pg/text@1']['output'];
    tasks: public_AnyTask[];
    readonly [RelationKeys]?: 'tasks';
  };
  export type public_Regular = {
    accountId: CodecTypes['pg/int4@1']['output'] | null;
    email: CodecTypes['pg/text@1']['output'];
    id: CodecTypes['pg/int4@1']['output'];
    kind: 'regular';
    name: CodecTypes['pg/text@1']['output'];
    plan: CodecTypes['pg/text@1']['output'] | null;
    account: public_Account | null;
    readonly [RelationKeys]?: 'account';
  };
  export type public_Task = {
    id: CodecTypes['pg/int4@1']['output'];
    projectId: CodecTypes['pg/int4@1']['output'] | null;
    reporterId: CodecTypes['pg/int4@1']['output'] | null;
    title: CodecTypes['pg/text@1']['output'];
    type: 'bug' | 'epic' | 'feature';
    comments: public_TaskComment[];
    project: public_Project | null;
    reporter: public_Person | null;
    readonly [RelationKeys]?: 'comments' | 'project' | 'reporter';
  };
  export type public_TaskComment = {
    body: CodecTypes['pg/text@1']['output'];
    id: CodecTypes['pg/int4@1']['output'];
    taskId: CodecTypes['pg/int4@1']['output'] | null;
    task: public_AnyTask | null;
    readonly [RelationKeys]?: 'task';
  };
  export type public_Ticket = {
    id: CodecTypes['pg/int4@1']['output'];
    ownerId: CodecTypes['pg/int4@1']['output'] | null;
    subject: CodecTypes['pg/text@1']['output'];
    owner: public_AnyUser | null;
    readonly [RelationKeys]?: 'owner';
  };
  export type public_User = {
    accountId: CodecTypes['pg/int4@1']['output'] | null;
    email: CodecTypes['pg/text@1']['output'];
    id: CodecTypes['pg/int4@1']['output'];
    kind: 'admin' | 'regular';
    name: CodecTypes['pg/text@1']['output'];
    account: public_Account | null;
    readonly [RelationKeys]?: 'account';
  };
  export type public_AnyTask = public_Bug | public_Epic | public_Feature;
  export type public_AnyUser = public_Admin | public_Regular;
}
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;